#172 - seperate user and auth operations
