+++ pages/admin/users/index.tsx
@@ -3,8 +3,8 @@ import { GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
 import { dehydrate, QueryClient } from 'react-query';
 import moment from 'moment';
-import { getUsers } from '../../../lib/api/auth';
-import { useDeleteUser, useGetUsers } from '../../../lib/operations/auth';
+import { getUsers } from '../../../lib/api/user';
+import { useDeleteUser, useGetUsers } from '../../../lib/operations/user';
 import { Entries } from '../../../components/Admin';
 import AddButton from '../../../components/Admin/Table/AddButton';
 import DeleteButton from '../../../components/Admin/Table/DeleteButton';
